+++ client/src/components/Topbar/MenuBullhorn.vue
@@ -1,13 +1,13 @@
 <template>
   <nav-item-dropdown
     :no-caret="!displayArrow"
-    tooltip="Aktionen"
+    :tooltip="$i18n('menu.tooltips.activities')"
     right
   >
     <template slot="button-content">
       <i class="fas fa-bullhorn" />
       <span v-if="displayText">
-        Aktionen
+        {{ $i18n('menu.tooltips.activities') }}
       </span>
     </template>
     <h3 class="dropdown-header">