+++ client/src/components/Topbar/MenuGroups.vue
@@ -1,8 +1,8 @@
 <template>
   <div>
     <nav-item-dropdown
+      :tooltip="$i18n('menu.tooltips.your_groups')"
       v-if="workingGroups.length"
-      tooltip="Deine Gruppen"
       no-caret
     >
       <template slot="button-content">
@@ -80,18 +80,20 @@
         role="menuitem"
         class="dropdown-item"
       >
-        <small><i class="fas fa-users" /> Gruppenübersicht</small>
+        <small><i class="fas fa-users" /> {{ $i18n('menu.tooltips.groups') }}</small>
       </a>
     </nav-item-dropdown>
     <li
+      :title="$i18n('menu.tooltips.groups')"
+      v-b-tooltip
       v-else
       class="nav-item"
     >
       <a
         v-b-tooltip
+        :title="$i18n('menu.tooltips.groups')"
         :href="$url('workingGroups')"
         class="nav-link"
-        title="Gruppenübersicht"
       >
         <i class="fas fa-users" />
       </a>